+++ src/main/java/frc/robot/commands/RotatePivotCommand.java
@@ -7,7 +7,7 @@ package frc.robot.commands;
 import java.util.function.DoubleSupplier;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Pivot;
+import frc.robot.subsystems.pivot.Pivot;
 
 public class RotatePivotCommand extends Command {
   private final Pivot pivot;
