+++ src/main/java/frc/robot/commands/SpinIntakeCommand.java
@@ -2,7 +2,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.intake.Intake;
 
 public class SpinIntakeCommand extends Command {
   private final Intake intake;
