+++ src/main/java/frc/robot/commands/SpinShooterCommand.java
@@ -5,7 +5,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Shooter;
+import frc.robot.subsystems.shooter.Shooter;
 
 public class SpinShooterCommand extends Command {
   private final Shooter shooter;
