+++ src/main/java/frc/robot/commands/StopClimberCommand.java
@@ -5,7 +5,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Climber;
+import frc.robot.subsystems.climber.Climber;
 
 public class StopClimberCommand extends Command {
   public final Climber climber;
