+++ src/main/java/frc/robot/commands/StopIntakeCommand.java
@@ -5,7 +5,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.intake.Intake;
 
 public class StopIntakeCommand extends Command {
   private final Intake intake;
