+++ src/main/java/frc/robot/commands/StopPivotCommand.java
@@ -5,7 +5,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Pivot;
+import frc.robot.subsystems.pivot.Pivot;
 
 public class StopPivotCommand extends Command {
   private final Pivot pivot;
