+++ src/main/java/frc/robot/commands/StopShooterCommand.java
@@ -5,7 +5,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Shooter;
+import frc.robot.subsystems.shooter.Shooter;
 
 public class StopShooterCommand extends Command {
   private final Shooter shooter;
