+++ src/main/java/frc/robot/commands/SwerveDriveCommand.java
@@ -15,7 +15,7 @@ import edu.wpi.first.math.kinematics.ChassisSpeeds;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.IOConstants;
-import frc.robot.subsystems.DriveBase;
+import frc.robot.subsystems.drive.DriveBase;
 
 /** A command to drive a swerve robot. */
 public class SwerveDriveCommand extends Command {
