+++ src/main/java/frc/robot/subsystems/climber/ClimberIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.climber;
 
 import org.littletonrobotics.junction.AutoLog;
