+++ src/main/java/frc/robot/subsystems/drive/DriveBase.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems;
+package frc.robot.subsystems.drive;
 import java.util.Optional;
 
 import org.littletonrobotics.junction.Logger;
@@ -29,6 +29,7 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Robot;
 import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.RobotConstants;
+import frc.robot.subsystems.vision.PhotonVision;
 
 public class DriveBase extends SubsystemBase {
   /** Creates a new SwerveDriveBase. */
