+++ src/main/java/frc/robot/subsystems/drive/SwerveModuleReal.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems;
+package frc.robot.subsystems.drive;
 
 import java.util.Arrays;
 
@@ -20,6 +20,8 @@ import edu.wpi.first.wpilibj.Timer;
 import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.SwerveConstants;
 import stl.math.LobstahMath;
+import stl.motorcontrol.MonitoredSparkMax;
+import stl.motorcontrol.TemperatureMonitor;
 
 public class SwerveModuleReal implements SwerveModuleIO {
   private final MonitoredSparkMax angleMotor;
