+++ src/main/java/frc/robot/subsystems/intake/IntakeIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.intake;
 
 import org.littletonrobotics.junction.AutoLog;
