+++ src/main/java/frc/robot/subsystems/intake/IntakeSparkMax.java
@@ -1,9 +1,12 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.intake;
 
 import java.util.Arrays;
 
 import com.revrobotics.CANSparkMax.IdleMode;
 import com.revrobotics.CANSparkMaxLowLevel.MotorType;
+
+import stl.motorcontrol.MonitoredSparkMax;
+import stl.motorcontrol.TemperatureMonitor;
 
 public class IntakeSparkMax implements IntakeIO {
     private final MonitoredSparkMax intakeMotor;
