+++ src/main/java/frc/robot/subsystems/shooter/Shooter.java
@@ -2,9 +2,11 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems;
+package frc.robot.subsystems.shooter;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import stl.motorcontrol.MonitoredSparkMax;
+import stl.motorcontrol.TemperatureMonitor;
 
 import java.util.Arrays;
