+++ src/main/java/frc/robot/subsystems/vision/PhotonVision.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.vision;
 
 import java.util.List;
 import java.util.Optional;
