+++ src/main/java/frc/robot/subsystems/vision/PhotonVisionIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.vision;
 
 import java.util.ArrayList;
 import java.util.List;
