+++ src/main/java/stl/motorcontrol/TemperatureMonitor.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems;
+package stl.motorcontrol;
 
 import java.util.List;
 
@@ -34,7 +34,7 @@ public class TemperatureMonitor {
     private boolean disabled = false;
     private List<Monitorable> motors;
 
-    TemperatureMonitor(List<Monitorable> motors) {
+    public TemperatureMonitor(List<Monitorable> motors) {
         this.motors = motors;
     }
